Obg 655: Tests split happy/unhappy
