+++ opba-protocols/xs2a-protocol-tests/xs2a-bdd-wiremock/src/test/java/de/adorsys/opba/protocol/xs2a/tests/e2e/wiremock/WiremockConsentE2EXs2aProtocolTest.java
@@ -10,7 +10,6 @@ import de.adorsys.opba.protocol.xs2a.tests.e2e.wiremock.mocks.WiremockAccountInf
 import de.adorsys.opba.protocol.xs2a.tests.e2e.wiremock.mocks.Xs2aProtocolApplication;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,7 +20,6 @@ import org.springframework.test.context.jdbc.Sql;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.nio.file.Path;
 import java.util.UUID;
 
 import static de.adorsys.opba.protocol.xs2a.tests.TestProfiles.MOCKED_SANDBOX;
@@ -112,7 +110,7 @@ class WiremockConsentE2EXs2aProtocolTest extends SpringScenarioTest<MockServers,
                 .open_banking_has_consent_for_anton_brueckner_transaction_list()
                 .fintech_calls_consent_activation_for_current_authorization_id()
                 .open_banking_can_read_anton_brueckner_transactions_data_using_consent_bound_to_service_session(
-                    ANTON_BRUECKNER_RESOURCE_ID, DATE_FROM, DATE_TO, BOTH_BOOKING
+                        ANTON_BRUECKNER_RESOURCE_ID, DATE_FROM, DATE_TO, BOTH_BOOKING
                 );
     }
 
@@ -192,157 +190,6 @@ class WiremockConsentE2EXs2aProtocolTest extends SpringScenarioTest<MockServers,
                 .user_max_musterman_selected_sca_challenge_type_email1_to_embedded_authorization()
                 .and()
                 .user_max_musterman_provided_sca_challenge_result_to_embedded_authorization_and_sees_redirect_to_fintech_ok();
-        then()
-                .open_banking_has_consent_for_max_musterman_transaction_list()
-                .fintech_calls_consent_activation_for_current_authorization_id()
-                .open_banking_can_read_max_musterman_transactions_data_using_consent_bound_to_service_session(
-                        MAX_MUSTERMAN_RESOURCE_ID, DATE_FROM, DATE_TO, BOTH_BOOKING
-                );
-    }
-
-    @ParameterizedTest
-    @EnumSource(Approach.class)
-    void testAccountsListWithOnceWrongPasswordThenOkWithConsentUsingEmbedded(Approach expectedApproach) {
-        given()
-                .embedded_mock_of_sandbox_for_max_musterman_accounts_running()
-                .preferred_sca_approach_selected_for_all_banks_in_opba(expectedApproach)
-                .rest_assured_points_to_opba_server()
-                .user_registered_in_opba_with_credentials(OPBA_LOGIN, OPBA_PASSWORD);
-
-        when()
-                .fintech_calls_list_accounts_for_max_musterman()
-                .and()
-                .user_logged_in_into_opba_as_opba_user_with_credentials_using_fintech_supplied_url(OPBA_LOGIN, OPBA_PASSWORD)
-                .and()
-                .user_max_musterman_provided_initial_parameters_to_list_accounts_all_accounts_consent()
-                .and()
-                .user_max_musterman_provided_wrong_password_to_embedded_authorization_and_stays_on_password_page()
-                .and()
-                .user_max_musterman_provided_correct_password_after_wrong_to_embedded_authorization()
-                .and()
-                .user_max_musterman_selected_sca_challenge_type_email2_to_embedded_authorization()
-                .and()
-                .user_max_musterman_provided_sca_challenge_result_to_embedded_authorization_and_sees_redirect_to_fintech_ok();
-        then()
-                .open_banking_has_consent_for_max_musterman_account_list()
-                .fintech_calls_consent_activation_for_current_authorization_id()
-                .open_banking_can_read_max_musterman_account_data_using_consent_bound_to_service_session();
-    }
-
-    @ParameterizedTest
-    @EnumSource(Approach.class)
-    void testTransactionsListWithOnceWrongPasswordThenOkWithConsentUsingEmbedded(Approach expectedApproach) {
-        given()
-                .embedded_mock_of_sandbox_for_max_musterman_transactions_running()
-                .preferred_sca_approach_selected_for_all_banks_in_opba(expectedApproach)
-                .rest_assured_points_to_opba_server()
-                .user_registered_in_opba_with_credentials(OPBA_LOGIN, OPBA_PASSWORD);
-
-        when()
-                .fintech_calls_list_transactions_for_max_musterman()
-                .and()
-                .user_logged_in_into_opba_as_opba_user_with_credentials_using_fintech_supplied_url(OPBA_LOGIN, OPBA_PASSWORD)
-                .and()
-                .user_max_musterman_provided_initial_parameters_to_list_transactions_with_single_account_consent()
-                .and()
-                .user_max_musterman_provided_wrong_password_to_embedded_authorization_and_stays_on_password_page()
-                .and()
-                .user_max_musterman_provided_correct_password_after_wrong_to_embedded_authorization()
-                .and()
-                .user_max_musterman_selected_sca_challenge_type_email1_to_embedded_authorization()
-                .and()
-                .user_max_musterman_provided_sca_challenge_result_to_embedded_authorization_and_sees_redirect_to_fintech_ok();
-        then()
-                .open_banking_has_consent_for_max_musterman_transaction_list()
-                .fintech_calls_consent_activation_for_current_authorization_id()
-                .open_banking_can_read_max_musterman_transactions_data_using_consent_bound_to_service_session(
-                        MAX_MUSTERMAN_RESOURCE_ID, DATE_FROM, DATE_TO, BOTH_BOOKING
-                );
-    }
-
-    @ParameterizedTest
-    @EnumSource(Approach.class)
-    void testTransactionsListWithOnceWrongIbanThenOkWithConsentUsingEmbedded(Approach approach) {
-        given()
-                .embedded_mock_of_sandbox_for_max_musterman_transactions_running()
-                .preferred_sca_approach_selected_for_all_banks_in_opba(approach)
-                .rest_assured_points_to_opba_server()
-                .user_registered_in_opba_with_credentials(OPBA_LOGIN, OPBA_PASSWORD);
-
-        when()
-                .fintech_calls_list_transactions_for_max_musterman()
-                .and()
-                .user_logged_in_into_opba_as_opba_user_with_credentials_using_fintech_supplied_url(OPBA_LOGIN, OPBA_PASSWORD)
-                .and()
-                .user_max_musterman_provided_initial_parameters_with_wrong_iban_to_list_transactions_with_single_account_consent()
-                .and()
-                .user_max_musterman_provided_initial_parameters_with_correct_iban_to_list_transactions_with_single_account_consent()
-                .and()
-                .user_max_musterman_provided_password_to_embedded_authorization()
-                .and()
-                .user_max_musterman_selected_sca_challenge_type_email1_to_embedded_authorization()
-                .and()
-                .user_max_musterman_provided_sca_challenge_result_to_embedded_authorization_and_sees_redirect_to_fintech_ok();
-        then()
-                .open_banking_has_consent_for_max_musterman_transaction_list()
-                .fintech_calls_consent_activation_for_current_authorization_id()
-                .open_banking_can_read_max_musterman_transactions_data_using_consent_bound_to_service_session(
-                        MAX_MUSTERMAN_RESOURCE_ID, DATE_FROM, DATE_TO, BOTH_BOOKING
-                );
-    }
-
-    @ParameterizedTest
-    @EnumSource(Approach.class)
-    void testAccountsListWithOnceWrongScaChallengeThenOkWithConsentUsingEmbedded(Approach expectedApproach) {
-        given()
-                .embedded_mock_of_sandbox_for_max_musterman_accounts_running()
-                .preferred_sca_approach_selected_for_all_banks_in_opba(expectedApproach)
-                .rest_assured_points_to_opba_server()
-                .user_registered_in_opba_with_credentials(OPBA_LOGIN, OPBA_PASSWORD);
-
-        when()
-                .fintech_calls_list_accounts_for_max_musterman()
-                .and()
-                .user_logged_in_into_opba_as_opba_user_with_credentials_using_fintech_supplied_url(OPBA_LOGIN, OPBA_PASSWORD)
-                .and()
-                .user_max_musterman_provided_initial_parameters_to_list_accounts_all_accounts_consent()
-                .and()
-                .user_max_musterman_provided_password_to_embedded_authorization()
-                .and()
-                .user_max_musterman_selected_sca_challenge_type_email2_to_embedded_authorization()
-                .and()
-                .user_max_musterman_provided_wrong_sca_challenge_result_to_embedded_authorization_and_stays_on_sca_page()
-                .and()
-                .user_max_musterman_provided_correct_sca_challenge_result_after_wrong_to_embedded_authorization_and_sees_redirect_to_fintech_ok();
-        then()
-                .open_banking_has_consent_for_max_musterman_account_list()
-                .fintech_calls_consent_activation_for_current_authorization_id()
-                .open_banking_can_read_max_musterman_account_data_using_consent_bound_to_service_session();
-    }
-
-    @ParameterizedTest
-    @EnumSource(Approach.class)
-    void testTransactionsListWithOnceWrongScaChallengeThenOkWithConsentUsingEmbedded(Approach expectedApproach) {
-        given()
-                .embedded_mock_of_sandbox_for_max_musterman_transactions_running()
-                .preferred_sca_approach_selected_for_all_banks_in_opba(expectedApproach)
-                .rest_assured_points_to_opba_server()
-                .user_registered_in_opba_with_credentials(OPBA_LOGIN, OPBA_PASSWORD);
-
-        when()
-                .fintech_calls_list_transactions_for_max_musterman()
-                .and()
-                .user_logged_in_into_opba_as_opba_user_with_credentials_using_fintech_supplied_url(OPBA_LOGIN, OPBA_PASSWORD)
-                .and()
-                .user_max_musterman_provided_initial_parameters_to_list_transactions_with_single_account_consent()
-                .and()
-                .user_max_musterman_provided_password_to_embedded_authorization()
-                .and()
-                .user_max_musterman_selected_sca_challenge_type_email1_to_embedded_authorization()
-                .and()
-                .user_max_musterman_provided_wrong_sca_challenge_result_to_embedded_authorization_and_stays_on_sca_page()
-                .and()
-                .user_max_musterman_provided_correct_sca_challenge_result_after_wrong_to_embedded_authorization_and_sees_redirect_to_fintech_ok();
         then()
                 .open_banking_has_consent_for_max_musterman_transaction_list()
                 .fintech_calls_consent_activation_for_current_authorization_id()
@@ -471,24 +318,6 @@ class WiremockConsentE2EXs2aProtocolTest extends SpringScenarioTest<MockServers,
                 .open_banking_can_read_anton_brueckner_account_data_using_consent_bound_to_service_session();
     }
 
-    @Test
-    void testAccountsListWithConsentUsingRedirectWithWrongIbans() {
-        given()
-                .redirect_mock_of_sandbox_for_anton_brueckner_accounts_running()
-                .preferred_sca_approach_selected_for_all_banks_in_opba(Approach.REDIRECT)
-                .rest_assured_points_to_opba_server()
-                .user_registered_in_opba_with_credentials(OPBA_LOGIN, OPBA_PASSWORD);
-
-        when()
-                .fintech_calls_list_accounts_for_anton_brueckner()
-                .and()
-                .user_logged_in_into_opba_as_opba_user_with_credentials_using_fintech_supplied_url(OPBA_LOGIN, OPBA_PASSWORD)
-                .and()
-                .user_anton_brueckner_provided_initial_parameters_to_list_accounts_with_wrong_ibans()
-                .and()
-                .got_503_http_error();
-    }
-
     @ParameterizedTest
     @EnumSource(Approach.class)
     void testTransactionsListWithConsentUsingEmbeddedForUnknownUser(Approach expectedApproach) {
@@ -518,75 +347,6 @@ class WiremockConsentE2EXs2aProtocolTest extends SpringScenarioTest<MockServers,
                 .open_banking_can_read_max_musterman_transactions_data_using_consent_bound_to_service_session(
                         MAX_MUSTERMAN_RESOURCE_ID, DATE_FROM, DATE_TO, BOTH_BOOKING
                 );
-    }
-
-    @Test
-    void testAccountsListWithConsentUsingRedirectWithoutIpAddress() {
-        given()
-                .redirect_mock_of_sandbox_for_anton_brueckner_accounts_running()
-                .preferred_sca_approach_selected_for_all_banks_in_opba(Approach.REDIRECT)
-                .rest_assured_points_to_opba_server()
-                .user_registered_in_opba_with_credentials(OPBA_LOGIN, OPBA_PASSWORD);
-        when()
-                .fintech_calls_list_accounts_for_anton_brueckner_no_ip_address()
-                .user_logged_in_into_opba_as_opba_user_with_credentials_using_fintech_supplied_url(OPBA_LOGIN, OPBA_PASSWORD);
-
-        then()
-                .user_anton_brueckner_provided_initial_parameters_to_list_accounts_with_all_accounts_consent_and_gets_202();
-    }
-
-    @ParameterizedTest
-    @EnumSource(Approach.class)
-    void testAccountsListWithConsentWithMissingPsuId(Approach expectedApproach, @TempDir Path tempDir) {
-        given()
-                .redirect_mock_of_sandbox_nopsu_for_anton_brueckner_accounts_running(tempDir)
-                .ignore_validation_rules_table_contains_field_psu_id()
-                .preferred_sca_approach_selected_for_all_banks_in_opba(expectedApproach)
-                .rest_assured_points_to_opba_server()
-                .user_registered_in_opba_with_credentials(OPBA_LOGIN, OPBA_PASSWORD);
-        when()
-                .fintech_calls_list_accounts_for_anton_brueckner()
-                .and()
-                .user_logged_in_into_opba_as_opba_user_with_credentials_using_fintech_supplied_url(OPBA_LOGIN, OPBA_PASSWORD)
-                .and()
-                .user_anton_brueckner_provided_initial_parameters_to_list_accounts_with_all_accounts_consent_without_psu_id()
-                .and()
-                .user_anton_brueckner_sees_that_he_needs_to_be_redirected_to_aspsp_and_redirects_to_aspsp()
-                .and()
-                .open_banking_redirect_from_aspsp_ok_webhook_called_for_api_test();
-        then()
-                .open_banking_has_consent_for_anton_brueckner_account_list()
-                .fintech_calls_consent_activation_for_current_authorization_id()
-                .open_banking_can_read_anton_brueckner_account_data_using_consent_bound_to_service_session();
-    }
-
-    @ParameterizedTest
-    @EnumSource(Approach.class)
-    void testAccountsListWithConsentUsingEmbeddedWithMissingIpAddressWhenItIsMandatoryByDefault(Approach expectedApproach) {
-        given()
-                .embedded_mock_of_sandbox_for_max_musterman_accounts_running()
-                .preferred_sca_approach_selected_for_all_banks_in_opba(expectedApproach)
-                .rest_assured_points_to_opba_server()
-                .user_registered_in_opba_with_credentials(OPBA_LOGIN, OPBA_PASSWORD);
-
-        when()
-                .fintech_calls_list_accounts_for_max_musterman_missing_ip_address()
-                .and()
-                .user_logged_in_into_opba_as_opba_user_with_credentials_using_fintech_supplied_url(OPBA_LOGIN, OPBA_PASSWORD)
-                .and()
-                .fintech_calls_get_consent_auth_state_to_read_violations_with_missing_ip_address()
-                .and()
-                .user_max_musterman_provided_initial_parameters_with_ip_address_to_list_accounts_all_accounts_consent()
-                .and()
-                .user_max_musterman_provided_password_to_embedded_authorization()
-                .and()
-                .user_max_musterman_selected_sca_challenge_type_email2_to_embedded_authorization()
-                .and()
-                .user_max_musterman_provided_sca_challenge_result_to_embedded_authorization_and_sees_redirect_to_fintech_ok();
-        then()
-                .open_banking_has_consent_for_max_musterman_account_list()
-                .fintech_calls_consent_activation_for_current_authorization_id()
-                .open_banking_can_read_max_musterman_account_data_using_consent_bound_to_service_session();
     }
 
     @ParameterizedTest
@@ -651,36 +411,6 @@ class WiremockConsentE2EXs2aProtocolTest extends SpringScenarioTest<MockServers,
 
     @ParameterizedTest
     @EnumSource(Approach.class)
-    void testAccountsListWithConsentUsingEmbeddedWithPsuIpPortWhenItIsOptionalInDb(Approach expectedApproach) {
-        given()
-                .embedded_mock_of_sandbox_for_max_musterman_accounts_running()
-                .ignore_validation_rules_table_ignore_missing_psu_ip_port()
-                .preferred_sca_approach_selected_for_all_banks_in_opba(expectedApproach)
-                .rest_assured_points_to_opba_server()
-                .user_registered_in_opba_with_credentials(OPBA_LOGIN, OPBA_PASSWORD);
-
-        when()
-                .fintech_calls_list_accounts_for_max_musterman()
-                .and()
-                .user_logged_in_into_opba_as_opba_user_with_credentials_using_fintech_supplied_url(OPBA_LOGIN, OPBA_PASSWORD)
-                .and()
-                .fintech_calls_get_consent_auth_state_to_read_violations_without_missing_psu_ip_port()
-                .and()
-                .user_max_musterman_provided_initial_parameters_to_list_accounts_all_accounts_consent()
-                .and()
-                .user_max_musterman_provided_password_to_embedded_authorization()
-                .and()
-                .user_max_musterman_selected_sca_challenge_type_email2_to_embedded_authorization()
-                .and()
-                .user_max_musterman_provided_sca_challenge_result_to_embedded_authorization_and_sees_redirect_to_fintech_ok();
-        then()
-                .open_banking_has_consent_for_max_musterman_account_list()
-                .fintech_calls_consent_activation_for_current_authorization_id()
-                .open_banking_can_read_max_musterman_account_data_using_consent_bound_to_service_session();
-    }
-
-    @ParameterizedTest
-    @EnumSource(Approach.class)
     void testAccountsListWithConsentUsingRedirectWithCookieValidation(Approach expectedApproach) {
         given()
                 .redirect_mock_of_sandbox_for_anton_brueckner_accounts_running()
@@ -708,5 +438,64 @@ class WiremockConsentE2EXs2aProtocolTest extends SpringScenarioTest<MockServers,
                 .open_banking_has_consent_for_anton_brueckner_account_list()
                 .fintech_calls_consent_activation_for_current_authorization_id()
                 .open_banking_can_read_anton_brueckner_account_data_using_consent_bound_to_service_session();
+    }
+
+    @ParameterizedTest
+    @EnumSource(Approach.class)
+    void testAccountsListWithConsentUsingEmbeddedWithMissingIpAddressWhenItIsMandatoryByDefault(Approach expectedApproach) {
+        given()
+                .embedded_mock_of_sandbox_for_max_musterman_accounts_running()
+                .preferred_sca_approach_selected_for_all_banks_in_opba(expectedApproach)
+                .rest_assured_points_to_opba_server()
+                .user_registered_in_opba_with_credentials(OPBA_LOGIN, OPBA_PASSWORD);
+
+        when()
+                .fintech_calls_list_accounts_for_max_musterman_missing_ip_address()
+                .and()
+                .user_logged_in_into_opba_as_opba_user_with_credentials_using_fintech_supplied_url(OPBA_LOGIN, OPBA_PASSWORD)
+                .and()
+                .fintech_calls_get_consent_auth_state_to_read_violations_with_missing_ip_address()
+                .and()
+                .user_max_musterman_provided_initial_parameters_with_ip_address_to_list_accounts_all_accounts_consent()
+                .and()
+                .user_max_musterman_provided_password_to_embedded_authorization()
+                .and()
+                .user_max_musterman_selected_sca_challenge_type_email2_to_embedded_authorization()
+                .and()
+                .user_max_musterman_provided_sca_challenge_result_to_embedded_authorization_and_sees_redirect_to_fintech_ok();
+        then()
+                .open_banking_has_consent_for_max_musterman_account_list()
+                .fintech_calls_consent_activation_for_current_authorization_id()
+                .open_banking_can_read_max_musterman_account_data_using_consent_bound_to_service_session();
+    }
+
+    @ParameterizedTest
+    @EnumSource(Approach.class)
+    void testAccountsListWithConsentUsingEmbeddedWithPsuIpPortWhenItIsOptionalInDb(Approach expectedApproach) {
+        given()
+                .embedded_mock_of_sandbox_for_max_musterman_accounts_running()
+                .ignore_validation_rules_table_ignore_missing_psu_ip_port()
+                .preferred_sca_approach_selected_for_all_banks_in_opba(expectedApproach)
+                .rest_assured_points_to_opba_server()
+                .user_registered_in_opba_with_credentials(OPBA_LOGIN, OPBA_PASSWORD);
+
+        when()
+                .fintech_calls_list_accounts_for_max_musterman()
+                .and()
+                .user_logged_in_into_opba_as_opba_user_with_credentials_using_fintech_supplied_url(OPBA_LOGIN, OPBA_PASSWORD)
+                .and()
+                .fintech_calls_get_consent_auth_state_to_read_violations_without_missing_psu_ip_port()
+                .and()
+                .user_max_musterman_provided_initial_parameters_to_list_accounts_all_accounts_consent()
+                .and()
+                .user_max_musterman_provided_password_to_embedded_authorization()
+                .and()
+                .user_max_musterman_selected_sca_challenge_type_email2_to_embedded_authorization()
+                .and()
+                .user_max_musterman_provided_sca_challenge_result_to_embedded_authorization_and_sees_redirect_to_fintech_ok();
+        then()
+                .open_banking_has_consent_for_max_musterman_account_list()
+                .fintech_calls_consent_activation_for_current_authorization_id()
+                .open_banking_can_read_max_musterman_account_data_using_consent_bound_to_service_session();
     }
 }
